project refactoring & make planner clear tasks include next 7 days
