+++ src/index.ts
@@ -5,9 +5,8 @@ import os from "os";
 import { LoggerService } from "./services/logger.service";
 
 // Routes
-import { router as samplesRouter } from "./routes/samples";
-import { router as suggestionsRouter } from "./routes/suggestions";
-import { router as plannerRouter } from "./routes/planner";
+import musicRouter from "./routes/music";
+import plannerRouter from "./routes/planner";
 
 // Variables
 const LOGGER = new LoggerService("index");
@@ -17,12 +16,8 @@ const initApp = () => {
   try {
     express()
       .use(express.json())
-      // Music Routes
-      .use("/v1/music", samplesRouter)
-      .use("/v1/music", suggestionsRouter)
-      // Planner Routes
+      .use("/v1/music", musicRouter)
       .use("/v1/planner", plannerRouter)
-      // Health
       .use("/health", (_, res: Response) =>
         res.status(200).send({
           errors: ERRORS,
@@ -35,7 +30,7 @@ const initApp = () => {
       })
       .listen(3000, () => {
         LOGGER.log(
-          `⚡️[server]: Server is running on PORT 3000 with architecture ${os.arch()}`
+          `⚡️[server]: Server is running on PORT 3000 on architecture ${os.arch()}`
         );
       });
   } catch (error: any) {
