--- src/services/util.service.ts
@@ -1,5 +1,0 @@
-export class UtilService {
-  static getTimestamp(): number {
-    return new Date().getTime();
-  }
-}
